refactor(lock): move lock stores out of cache handlers
- introduce cache lock store provider and dedicated File, Redis, and Predis lock stores
- keep cache handlers responsible only for exposing their lock store
- update LockManager to resolve lock support through cache lock store providers
- refresh memoized Redis and Predis lock stores after reconnect
- update Redis and Predis lock tests for provider-based stores and reconnect behavior
- document custom lock store extension expectations
