+++ system/Cache/Handlers/RedisHandler.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace CodeIgniter\Cache\Handlers;
 
 use CodeIgniter\Cache\LockStoreInterface;
+use CodeIgniter\Cache\LockStoreProvider;
+use CodeIgniter\Cache\LockStores\RedisLockStore;
 use CodeIgniter\Exceptions\CriticalError;
 use CodeIgniter\I18n\Time;
 use Config\Cache;
@@ -25,7 +27,7 @@ use RedisException;
  *
  * @see \CodeIgniter\Cache\Handlers\RedisHandlerTest
  */
-class RedisHandler extends BaseHandler implements LockStoreInterface
+class RedisHandler extends BaseHandler implements LockStoreProvider
 {
     /**
      * Default config
@@ -55,6 +57,8 @@ class RedisHandler extends BaseHandler implements LockStoreInterface
      */
     protected $redis;
 
+    private ?LockStoreInterface $lockStore = null;
+
     /**
      * Note: Use `CacheFactory::getHandler()` to instantiate.
      */
@@ -69,7 +73,8 @@ class RedisHandler extends BaseHandler implements LockStoreInterface
     {
         $config = $this->config;
 
-        $this->redis = new Redis();
+        $this->redis     = new Redis();
+        $this->lockStore = null;
 
         try {
             $funcConnection = isset($config['persistent']) && $config['persistent'] ? 'pconnect' : 'connect';
@@ -186,59 +191,6 @@ class RedisHandler extends BaseHandler implements LockStoreInterface
         return $this->increment($key, -$offset);
     }
 
-    public function acquireLock(string $key, string $owner, int $ttl): bool
-    {
-        $key = static::validateKey($key, $this->prefix);
-
-        return (bool) $this->redis->set($key, $owner, ['nx', 'ex' => $ttl]);
-    }
-
-    public function releaseLock(string $key, string $owner): bool
-    {
-        $key = static::validateKey($key, $this->prefix);
-
-        $script = <<<'LUA'
-            if redis.call("get", KEYS[1]) == ARGV[1] then
-                return redis.call("del", KEYS[1])
-            end
-
-            return 0
-            LUA;
-
-        return (int) $this->redis->eval($script, [$key, $owner], 1) === 1;
-    }
-
-    public function forceReleaseLock(string $key): bool
-    {
-        $key     = static::validateKey($key, $this->prefix);
-        $deleted = $this->redis->del($key);
-
-        return is_int($deleted) && $deleted >= 0;
-    }
-
-    public function refreshLock(string $key, string $owner, int $ttl): bool
-    {
-        $key = static::validateKey($key, $this->prefix);
-
-        $script = <<<'LUA'
-            if redis.call("get", KEYS[1]) == ARGV[1] then
-                return redis.call("expire", KEYS[1], ARGV[2])
-            end
-
-            return 0
-            LUA;
-
-        return (int) $this->redis->eval($script, [$key, $owner, $ttl], 1) === 1;
-    }
-
-    public function getLockOwner(string $key): ?string
-    {
-        $key   = static::validateKey($key, $this->prefix);
-        $owner = $this->redis->get($key);
-
-        return is_string($owner) ? $owner : null;
-    }
-
     public function clean(): bool
     {
         return $this->redis->flushDB();
@@ -271,6 +223,13 @@ class RedisHandler extends BaseHandler implements LockStoreInterface
     public function isSupported(): bool
     {
         return extension_loaded('redis');
+    }
+
+    public function lockStore(): LockStoreInterface
+    {
+        assert($this->redis instanceof Redis);
+
+        return $this->lockStore ??= new RedisLockStore($this->redis, $this->prefix);
     }
 
     public function ping(): bool
